fix: pudding modal route no longer renders details page
I resisted following this approach from the beginning because it makes
it less clear which routes have been migrated or not. Also, I was hoping
for a more streamlined process to migrating routes, meaning the same PR
that would add a route to the framework router would also modify the
component to use `params`, `loaderData` and `actionData` at least, even
if not SSR for now.

At the end it's not _that_ big of a deal, but in a corporate environment
this have to be accorded with the team and a pattern must be defined to
avoid confusion and non-migrated pages getting left in limbo.